backend/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.5
django-filter==25.1
djangorestframework==3.15.2
drf-yasg==1.21.8
greenlet==3.1.1
inflection==0.5.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
SQLAlchemy==2.0.37
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
joblib
matplotlib
seaborn
scikit-learn
django-cors-headers
Pillow

